feat: Implement "Selection JSON Specification"
Implement the optional Selection JSON Specification. This feature allows, using a "Selection Scenario" (generally retrieved once from the server), to perform the card detection locally on the Terminal endpoint and transmit the result of this Selection Scenario to the Keyple server on the first http contact.
This feature saves (at least) one HTTP round trip, which is significant for a "typical" keyple session that uses 3 or 4 HTTP round trips. Hence it's highly recommended to use it!

The documentation is available here: https://keyple.org/user-guides/non-keyple-client/selection-json-specification/