fix: better error handling
When an error occurs locally (on the device), we need to properly finish the job and that may lead to new errors popping up. It's more likely that only the first error is relevant and actionable for the user. Change the flow to save the first "local" error, and return it to the user after properly finishing the transaction with the server (even if network or other errors occurs).