fix: return both local error and server outputData
When a transaction fails locally, we store the local error and finish the transaction with the server. We return both the local error (which is probably of the most relevance to the user) and the server generated outputData, so that the app developper has all he needed info to proceed.